+++ plugin.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 """
-<plugin key="MySkodaAPI" name="MySkoda API Integration" author="Jan Reimen" version="0.4.2" externallink="https://github.com/janreimen/Domoticz-MySkodaAPI">
+<plugin key="MySkodaAPI" name="MySkoda API Integration" author="Jan Reimen" version="0.4.3" externallink="https://github.com/janreimen/Domoticz-MySkodaAPI">
   <description>
     <h2>MySkoda API Integration</h2><br/>
     Read-only integration with the official Škoda MySkoda Public API.<br/>